requirements.txt
asgiref==3.7.2
Django==5.0.1
django-filter==23.5
django-widget-tweaks==1.5.0
pillow==10.2.0
sqlparse==0.4.4
